fix: preserve legacy volume urls and split page client
